--- Frontend/src/pages/Group/Group.module.scss
@@ -1,119 +1,0 @@
-html {
-  font-size: 62.5%; 
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.container {
-  display: flex;
-  flex-direction: column; 
-  gap: 2rem; 
-  padding: 2rem;
-  background-color: #fff;
-  color: #333;
-}
-
-h2 {
-  color: #333;
-  font-size: 3rem; 
-}
-
-.groupRow {
-  display: flex;
-  gap: 2rem;
-  flex-wrap: wrap; 
-}
-
-.groupCard {
-  position: relative; 
-  background-color: #e6f0fd;
-  padding: 1.5rem;
-  border-radius: 0.8rem;
-  text-align: center;
-  width: calc(33.33% - 5rem);
-  cursor: pointer;
-  transition: transform 0.2s;
-  font-size: 1.4rem;
-
-  h3{
-    margin-top: 1rem;
-    font-size: 2rem;
-    font-weight: bold;
-  }
-
-}
-
-.groupCard:hover {
-  transform: scale(1.05);
-
-}
-
-.groupCard:hover .removeButton {
-  display: block;
-}
-
-.groupImage {
-  width: 100%;
-  height: 15rem;
-  object-fit: cover;
-  border-radius: 0.5rem;
-}
-
-.joinButton {
-  margin-top: 1rem; 
-  width: 80%;
-  padding: 1rem; 
-  font-weight: bold;
-  font-size: 1.4rem;
-  background-color: #008cba; 
-  color: white;
-  border: none;
-  border-radius: 0.5rem;
-  cursor: pointer;
-
-}
-
-.joinButton:hover {
-  background-color: #005f6b;
-}
-
-.joinedButton {
-  margin-top: 1rem; 
-  width: 80%;
-  font-weight: bold;
-  padding: 1rem; 
-  font-size: 1.4rem;
-  background-color: #555; 
-  color: #ccc;
-  border: none;
-  border-radius: 0.5rem;
-  
-}
-
-.removeButton {
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-  background-color: #ccc;
-  border: none;
-  border-radius: 50%;
-  color: #fff;
-  font-size: 2.2rem; 
-  cursor: pointer;
-  display: none;
-}
-
-.removeButton:hover {
-  background-color: red;
-}
-
-
-
-
-.joinedGroups {
-  margin-bottom: 2rem;
-}
-
-
-.suggestedGroups {
-  margin-bottom: 2rem;
-}
